nerpa_controller: Skip unknown tables in push_outputs().
This code fully processes every delta that DDlog passes along, even if the
delta isn't for a table that will be pushed to the P4Runtime controller,
which meant that if there were any output tables that had types that the
code couldn't process, then it would panic.  It's not unusual to have
output tables that aren't meant for P4Runtime, so it seems bad to restrict
those tables to types that are suitable for P4Runtime.  This fixes the
problem by skipping those tables instead of processing them.

With this change, rustc noticed that there were some now-unused values, so
I restructured the code slightly to avoid that.