nerpa_controller, p4ext, snvs: make snvs work end-to-end  (#19)
* Configure primary when streaming digests.

This configures the primary controller when streaming digests. This
fixes previous reliability issues when streaming. Before, we only
set primary when first running the controller. Now, whenever a
streaming channel is established, it is configured as primary. That
guarantees that all notifications will be streamed over the channel.

* Process subcrates in p4info2ddlog.

This properly processes subcrates in p4info2ddlog. It makes sure
that generated code and dependencies properly import DDlog types.
This fixes potential runtime errors when translating digests into
DDlog types.

* nerpa_controller, p4ext: Support fields up to 64 bits wide.

This code assumed that fields were at most 16 bits wide, which is often
not the case.

* nerpa_controller: New command-line option --ddlog-record.

This makes it easier to see what's going on and to replay it with the
DDlog CLI.

Also, use the option in run-nerpa.sh.

* nerpa_controller: Drop some unused 'self' parameters.

There's no point in passing a 'self' parameter to these functions, because
they do not use it.

* nerpa_controller: Pick a default entry action when there's an obvious one.

When a P4 table only has one action that may be used in table entries, and
that action takes no parameters, there is no reason to make the programmer
state the action explicitly, so this commit does it automatically.

* nerpa_controller: Translate Some{x} and None from DDlog to P4Runtime.

Otherwise extract_record_value() will panic (currently) or mistranslate to
zero (previously).

This can't be part of extract_record_value() at the moment because that
function just returns a u64.  It's possible that that function should
return something more flexible.  Further development will probably show
which is better.

* p4ext: Tolerate omitted fields in building table entries.

Most P4Runtime match types allow the match field to be omitted, and in fact
require it to be omitted if the match field should not be matched.  This
code would require an error in this case.  This fixes the problem.

* nerpa_controller: Skip unknown tables in push_outputs().

This code fully processes every delta that DDlog passes along, even if the
delta isn't for a table that will be pushed to the P4Runtime controller,
which meant that if there were any output tables that had types that the
code couldn't process, then it would panic.  It's not unusual to have
output tables that aren't meant for P4Runtime, so it seems bad to restrict
those tables to types that are suitable for P4Runtime.  This fixes the
problem by skipping those tables instead of processing them.

With this change, rustc noticed that there were some now-unused values, so
I restructured the code slightly to avoid that.

* nerpa_controller: Push DDlog initial delta to switch.

DDlog code might produce some output data at initialization time, before
the controller adds any input data.  The controller was discarding these
initial outputs.  This code makes the controller instead push them to the
switch.

* snvs: Rework for demo 3.

Co-authored-by: Debnil Sur <dsur@vmware.com>