Meaningfully relate P4 and DDlog programs.
This modifies the P4 and DDlog programs to have a meaningful
relationship between the DDlog rules and the P4 tables. To do this,
it just takes some egress code from the snvs P4 program and writes
an accompanying DDlog rule. This lets us properly test p4info2ddlog
in the CI steps.