Add simple automatic testing infrastructure and snvs example.
Until now, this project didn't support automatic testing.  This
commit adds a little bit of infastructure for this through a
new "bmv2_packet" library that can communicate with the P4
behavioral model implementation, passing packets back and forth
each way.

To use this, it's necessary to invoke bmv2 with a special option
--packet-in that tells it to use nanomsg to simulate its
interfaces rather than to go through system interfaces like
veths.  This commit adds the "-s" option to the test-nerpa.sh
script to make it do that.  As a side effect, this means that
running test-nerpa.sh doesn't require superuser privilege
(because it's not trying to configure or use veths).

This commit also adds a program to the snvs example that runs
an automatic test against a running bmv2 started in this way.