scripts: Use OVSDB in a more Nerpa-specific fashion.
OVSDB is just a database, and you can use it to keep track of anything,
but its defaults all assume you're going to use it to configure Open
vSwitch: it listens on a socket in /var/run/openvswitch, its default
database is in /etc/openvswitch, and so on.  This doesn't make sense for
Nerpa, but the Nerpa scripts so far don't do a good job of distinguishing.

This commit makes Nerpa use OVSDB in its own way.  Instead of using the
system directories and databases, it creates its own files in Nerpa's
own directory using the application's own schema.  It also creates the
database at runtime rather than build time.

Finally, this just uses --remote when it starts ovsdb-server instead of
adding a remote afterward.