nerpa_controlplane, nerpa_controller: implement ARP (#24)
* Handle Nerpa programs without OVS Schema.

Correctly handles Nerpa programs that do not have an OVS Schema
provided.

* Add ARP P4 program.

This also edits p4info2ddlog to properly process the lpm match.

* Remove snvs-specific extern from ovsdb_client.

* Move multicast update into a separate function.

* Send packets from dataplane to controller in ARP.

* Parse packets from the dataplane in dp2ddlog and the controller.

This adds necessary functionality to parse packets in dataplane
messages in dp2ddlog and the controller. It generates DDlog input
relations corresponding to controller_packet_metadata in the P4info.

* Decomposes p4info2ddlog by P4info data type for readability.

* Successfully push an ARP output to switch.

* Process more complex action and field matches.

This correctly pushes more complex action and field matches to the
switch. Sending a packet to the data plane does not work yet. This
also includes necessary changes to the p4ext library.

* Remove dead code from p4ext library.

Changing the data transformation procedure from DDlog
to p4runtime resulted in significant dead code in the p4ext
library. This removes that dead code.

* Successfully send packet out to switch.

* Implement ARP in the DDlog program.

* Run clippy.

* Make requested documented changes.

This adds requested changes to documentation. The changes focus on
the expected format of DDlog output records in a Nerpa program.
A future PR will add documentation across the existing codebase.

* Add logging to the controller.

* Ignore OVS files from testing.

* Connect to the switch to send packets in the constructor.

* Stop removing DDlog crate per-build.

* Improve dataplane message handling.

This changes log granularity and early-returns empty messages.

Co-authored-by: Debnil Sur <dsur@vmware.com>