nerpa_controller: Trim leading 0-bytes off P4Runtime bytestrings.
P4Runtime uses bytestrings to represent arbitrary-length bit<> and int<>
fields.  These are supposed to be the minimum length for their content
regardless of the field that they go into.  The code here was in many
cases instead padding them out to the field width.  This caused some
confusion when I was implementing the other side of P4Runtime.  This commit
eliminates the padding, which also simplifies the code in places because
the desired bit-width doesn't have to be tracked or passed around.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>