ovsdb-sys: Rename 'ovs' and expand to cover more of OVS.
The ovsdb-sys crate wrapped the part of the OVS libraries needed to act as
an OVSDB client.  This commit wraps a larger fraction of the OVS libraries.
Thus, the name 'ovsdb-sys' is now not general enough, so this commit
renames it to just 'ovs' and puts the bindgen-generated unsafe bits into
an 'ovs::sys' module.

This commit updates all the code that used ovsdb_sys to instead use
ovs::sys.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>