Push output changes to multiple P4-enabled devices.
This rewrites the ControllerActor to support multiple switch
clients, representing multiple P4-enabled devices. We now pass
a Vector of SwitchClient structs to the actor, and it pushes
outputs to each of those clients. This does include some redundant
computation in converting the output relation to the appropriate
format for the switch.

Signed-off-by: Debnil Sur <dsur@vmware.com>