p4c-of: Fix OpenFlow field names and their prerequisites.
p4c-of was using header field names as OpenFlow names and header names as
their prerequisites.  This is a simplification that is often wrong.  This
commit adds explicit name annotations for field names (as @name) and
their prerequisites (as @of_prereq) and adapts of_model.p4 to use them.

Some of the details are still wrong, e.g. this makes access to TCP
headers assume a tcpv4 prerequisite.  I am not sure of the correct
solution.  It might be to have separate tcpv4 and tcpv6 headers.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>