p4c-of: Fix bitwise matches.
Bitwise matches in flows used the wrong syntax.  To match a 1-bit with
value 0x100 in reg1, it used the incorrect syntax "reg1/0x100=1".  The
correct syntax is "reg1=0x100/0x100", that is, "field=value/mask" in
which "value" is shifted to match up bitwise with "mask".  This commit
fixes the problem.

One contribution to the problem is that OF_Fieldname doesn't have any
provisions to keep track of a bitwise slice, which is necessary to be able
to handle expressions that match against them.  One way to do this would
be to handle OF_Slice with an inner OF_Fieldname.  But there's an existing
OF_Register that is suitable and already has slice support.  The only
problem is that it is specialized for registers name-wise and size-wise.
Thus, this commit adapts OF_Register so it can handle any field, and
switches uses of OF_Fieldname to OF_Register wherever that is appropriate
(which is for anything other than 'table=X', 'priority=X', prerequisites,
and action parameter names).

It probably makes sense to rename OF_Register to OF_Field and OF_Fieldname
to something else too (since it's never used for field matches anymore).
This commit doesn't do that.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>