p4c-of: Implement P4 fields that are slices of OF fields.
A lot of P4 fields map reasonably well to whole OpenFlow fields, but some
of them make more sense as a slice of an OpenFlow field.  The example of
the latter that comes to mind immediately are the sub-fields of
Open vSwitch's vlan_tci, which has VID and PCP subfields and a bit that
indicates whether a VLAN header is present.

This commit implements these "slice" fields as a new @of_slice annotation
and adopts it for the Vlan header in of_model.p4.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>