p4c-of: Correctly slice an OF_Register that has a friendlyName.
When OF_Register has a friendlyName, that indicates that a function whose
name is derived from it can be invoked to get its allocation, which
includes slice information.  If we take a further slice, then that
slice information is wrong.  (It can't be fixed by appending a second
slice operator, because Open vSwitch won't accept syntax like
"reg1[5..10][1..2]".)  Thus, we must drop the friendlyName when we
slice an OF_Register.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>