p4c-of: Combine multiple matches on a single field in OpenFlow output.
Open vSwitch doesn't accept multiple matches on a single field, even if
they are disjoint, like "field=0x34/0xff, field=0x1200/0xff00".  Rather,
multiple matches have to be combined, like "field=0x1234/0xffff".  p4c-of
didn't do this.  This commit implements such a transformation.

In the output for snvs.p4, this corrects the following:
    vlan_tci=${has_vlan << 12}/0x1000, vlan_tci=${vid}/0xfff
into:
    vlan_tci=${(has_vlan << 12) | vid}/0x1fff

As a side effect, this change changes the ordering of output fields in
the generated OpenFlow.  This shouldn't cause a visible change in behavior.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>