ofp4: Implement zero-extension for copying between fields.
Without this, an assignment in P4 from a narrower to a wider field, even
with a cast, will fail with an error when OVS parses it, because OVS
requires the width of both operands of 'move' to be the same.

This fixes two cases of zero-extension in snvs.dl that previously caused
an error when OVS parsed them.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>