ofp4: Update snvs.dl to match current output of p4c-of.
The idea of having snvs.dl in the tree is to make it clear what's changing
in the output of p4c-of on each commit.  (Ultimately this should be
replaced by proper tests, and that's the idea, and I'm working on it.)
However, it had gotten out of sync.  This updates it.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>