ofp4: Log actual listening address when the kernel picks for us.
When one specifies a port number of 0 to the kernel for listening on a
socket, it picks a free port number automatically.  This is useful for
testing, because if you specify any particular port it might already be
in use.  But if you specify port number 0, the test harness has to figure
out what port is really in use.  Looking in the log is one reasonable way,
so this commit causes ofp4 to log the port number if the kernel picked it.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>