p4c-of: Support constant entries in P4 tables in a more uniform manner.
Before this commit, constant entries in P4 were handled by adding special
constant Flow entries to the output DDlog.  This works OK but it's a
special case in how table entries are emitted, since ordinarily a generic
Flow gets its parameters from an input table.

This commit makes the special case for constant entries closer to the
general case.  Instead of generating a special Flow for each constant
entry, it generates the same generic Flow as in the general case, and then
adds an input relation with a built-in row for each constant entry.  This
allows more shared code between the two cases.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>