p4c-of: Support constant default actions in a more uniform manner.
Before this commit, constant table default actions in P4 were handled
by adding special constanf Flow entries to the output DDlog.  This works
OK but it's a special case in how actions are emitted, since ordinarily a
generic Flow gets its parameters from a DefaultAction table.

This commit makes the special case for constant default actions closer to
the general case.  Instead of generating a special Flow for each constant
default action, it generates the same generic Flow as in the general case,
and then adds an input relation with a built-in row for the constant
default action.  This allows more shared code between the two cases.

Signed-off-by: Ben Pfaff <bpfaff@vmware.com>